+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed the rows with cash positions in the holdings table
 - Fixed an issue with the date parsing in the historical market data editor of the admin control panel
 
 ## 1.300.0 - 2023-08-11